+++ pytest_helm_charts/flux/git_repository.py
@@ -4,7 +4,7 @@ from typing import Protocol, Optional, Any, List, Dict
 from pykube import HTTPClient
 
 from pytest_helm_charts.k8s.fixtures import NamespaceFactoryFunc
-from pytest_helm_charts.flux.utils import NamespacedFluxCR, FLUX_CR_READY_TIMEOUT_SEC, _flux_cr_ready
+from pytest_helm_charts.flux.utils import NamespacedFluxCR, FLUX_CR_READY_TIMEOUT_SEC, flux_cr_ready
 from pytest_helm_charts.utils import wait_for_objects_condition, inject_extra
 
 
@@ -150,7 +150,7 @@ def wait_for_git_repositories_to_be_ready(
         GitRepositoryCR,
         git_repo_names,
         git_repo_namespace,
-        _flux_cr_ready,
+        flux_cr_ready,
         timeout_sec,
         missing_ok,
     )